requirements.txt
torch==1.4.0
opencv-python==4.2.0.34
torchvision==1.5.0
tqdm==4.46.1
scipy==1.4.1
numpy==1.18.2
Pillow==7.1.2
scikit_learn==0.23.1
